Move mock data to separate file for edit event page tests
